pyproject.toml
[project]
name = "gcp-pipeline"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cloud-sql-python-connector[pg8000]>=1.20.1",
    "commitizen>=4.13.9",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
]

[project.scripts]
gcp-pipeline = "gcp_pipeline:main"

[build-system]
requires = ["uv_build>=0.9.7,<=0.10.11"]
build-backend = "uv_build"

[tool.ruff]
line-length = 120

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "semver"
version_provider = "uv"
